scripts/requirements.txt
Pillow==10.1.0
watchdog==3.0.0